requirements-macos.txt
numpy
pandas
tensorflow-macos==2.15.0
tensorflow-metal
keras==2.15.0
keras_applications >= 1.0.7
opencv-python>=4.1.0
mtcnn>=0.1.0
h5py
efficientnet
split_folders
matplotlib